Add files via upload
changed how demographic data collection (Age, Sex, Race, Ethnicity) are modeled  as outcomes of a DemographicDataCollection activity. The date of that activity maps to the SDTM DM.DMDTC variable. The benefit is the activity resource can in the future (if necessary) link to other information, such as the method of collection/derivation, etc. and the information can easily be updated with additional instances of the activity class.